update profile issue fixed, delete profile picture added

+++ src/main/java/org/ahmedukamel/eduai/service/user_management/UserManagementService.java
@@ -3,11 +3,10 @@ package org.ahmedukamel.eduai.service.user_management;
 import lombok.RequiredArgsConstructor;
 import org.ahmedukamel.eduai.constant.PathConstants;
 import org.ahmedukamel.eduai.dto.api.ApiResponse;
-import org.ahmedukamel.eduai.model.Event;
 import org.ahmedukamel.eduai.model.User;
 import org.ahmedukamel.eduai.repository.UserRepository;
 import org.ahmedukamel.eduai.saver.file.FileSaver;
-import org.ahmedukamel.eduai.service.db.DatabaseService;
+import org.ahmedukamel.eduai.util.context.ContextHolderUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -23,13 +22,15 @@ public class UserManagementService implements IUserManagementService {
     private final FileSaver fileSaver;
 
     @Override
-    public Object uploadProfilePicture(Long id, MultipartFile picture){
-        User user = DatabaseService.get(userRepository::findById, id, Event.class);
+    public Object uploadProfilePicture(MultipartFile picture){
+        User user = ContextHolderUtils.getUser();
 
         try {
-            Path oldFilePath = PathConstants.PROFILE_PHOTOS_PATH.resolve(user.getPicture());
+            if(user.getPicture() != null) {
+                Path oldFilePath = PathConstants.PROFILE_PHOTOS_PATH.resolve(user.getPicture());
+                Files.delete(oldFilePath);
+            }
 
-            Files.delete(oldFilePath);
 
             String filename = fileSaver.save(picture, PathConstants.PROFILE_PHOTOS_PATH);
 
@@ -48,5 +49,27 @@ public class UserManagementService implements IUserManagementService {
         String message = "User picture updated successfully.";
 
         return new ApiResponse(true, message, "");
+    }
+
+    @Override
+    public Object removeProfilePicture() {
+        User user = ContextHolderUtils.getUser();
+
+        if(user.getPicture() != null) {
+            try {
+                Files.delete(PathConstants.PROFILE_PHOTOS_PATH.resolve(user.getPicture()));
+            } catch (IOException exception) {
+                throw new RuntimeException("Failed delete file.", exception);
+            }
+        }
+
+
+        user.setPicture(null);
+        userRepository.save(user);
+
+        String message = "User profile picture deleted successfully.";
+
+        return new ApiResponse(true, message, "");
+
     }
 }
